==> Code/1-compile_phage_training_data.ipynb <==
# Notes:

This notebook is for compiling and processing the phage genomes that will be used as ground-truth benchmarks for training/testing a lifestyle prediction model. The notebook should be stepped through top to bottom and assumes that you have cloned this repository in it's entirety, including data folders with identical directory structures. No promises made on how it will run on Windows machines.

Of particular note is that this notebook is **very specific** to **this particular phage** training dataset and the code will need to be re-factored depending on the format (data tables, column names, etc.) of any potential future training datasets.

# Imports

import pandas as pd
from Bio import SeqIO, Entrez
import glob

import time
import os

import bacphlip

# Constants

base_dir = '../Data/phage_data_nmicro2017/'
annotations_file = base_dir + 'original_data/41564_2017_BFnmicrobiol2017112_MOESM2_ESM.csv'
phagesdb_fasta_file = base_dir + 'original_data/Actinobacteriophages-All.fasta'
fasta_dir = base_dir + 'phage_fasta_files/'
processed_df_outfile = base_dir + 'processed_benchmark_set.csv'

##############################################
#If you're re-running this dataframe with the intention of downloading or re-downloading
#genome data in bulk from NCBI please be sure to provide your email address for them
#If you have cloned the full repository for this project, all files should be accessible 
#and will not re-download by default making this superfluous
Entrez.email = 'ENTER YOUR EMAIL ADDRESS HERE'
##############################################

###Make sure that all the directories and file paths are valid
for directory in [base_dir, fasta_dir]:
    assert os.path.exists(directory)

for file_of_interest in [annotations_file, phagesdb_fasta_file]:
    assert os.path.exists(file_of_interest)

# Select phages of interest from the starting file

df = pd.read_csv(annotations_file)
print('Starting dataframe shape:', df.shape)
df = df[df['Temperate (empirical)'] != 'Unspecified']
print('Ground truth dataframe shape:', df.shape)
df.head()

df.columns

df['Host phylum'].value_counts()

df['Host class'].value_counts()

df['Host order'].value_counts()

# Scan through bulk fasta file from PhagesDB to isolate and write individual records of interest

Note that this file was retreived in March 2020, from https://phagesdb.org/data/

The relevant data file was listed under the heading:
"Download Multifasta of All Actinobacteriophage Genomes"

fastas = list(SeqIO.parse(phagesdb_fasta_file, 'fasta'))
print('Contains {} separate fasta sequences'.format(len(fastas)))
print()
print('########## Here is an example entry:')
print(fastas[0])

**Pattern matching data to pair the `csv` file entries with `fasta` records**

temp_df = df[df['Database source'] == 'Actinobacteriophage_785']
problem_cases = []
for index in temp_df.index[:]:
    found = []
    identifier = temp_df.loc[index]['Virus identifier used for the analysis'].split('_')[0].lower()
    for fasta in fastas:
        if identifier in fasta.description.lower().split(' '):
            found.append(fasta)
    if len(found) != 1:
        problem_cases.append(identifier)
        print(identifier, len(found))
        continue
    SeqIO.write(found[0], fasta_dir + '{}.fasta'.format(identifier), 'fasta')

###Cleaning up and dealing with some of those problem cases
for index in temp_df.index[:]:
    found = []
    identifier = temp_df.loc[index]['Virus identifier used for the analysis'].split('_')[0].lower()
    
    ########
    if identifier not in problem_cases:
        continue
    ########
    
    for fasta in fastas:
        if identifier in fasta.description.lower().replace(',', ' ').replace('_', ' ').split(' '):
            found.append(fasta)
    if len(found) != 1:
        problem_cases.append(identifier)
        print(identifier, len(found))
        continue
    SeqIO.write(found[0], fasta_dir + '{}.fasta'.format(identifier), 'fasta')

###Remove the problematic cases that for some reason or another I can't easily find
df = df[df['Virus identifier used for the analysis'].str.contains('schwabeltier', case=False) == False]
df = df[df['Virus identifier used for the analysis'].str.contains('theloniousmonk', case=False) == False]
print(df.shape)

**Write the new dataframe containing all of the actinobacteria phages that I was able  to find and for which we have lifestyle annotations**

df.to_csv(processed_df_outfile)

# Download RefSeq data from NCBI

This section will download full genome fasta files from NCBI and as such may take some time to run, and should probably be run at night time. In reality, the number is fairly small so it's not placing a heavy burden on NCBI servers but be conscious and aware of bulk downloading in this manner

temp_df = df[df['Database source'] == 'NCBI RefSeq']
print(temp_df.shape)

for index in temp_df.index[:]:
    print(index)
    accession = temp_df.loc[index]['RefSeq accession number']
    
    #####################################################################
    ###Don't download if I already have the file downloaded and ready to go
    if os.path.exists(fasta_dir + '{}.fasta'.format(accession)):
        continue
    #####################################################################

    handle = Entrez.efetch(db='nuccore', id=accession, rettype='fasta', retmode='text')
    text = handle.read()
    handle.close()
    with open(fasta_dir + '{}.fasta'.format(accession), 'w') as outfile:
        outfile.write(text)
    time.sleep(10)

**Ensuring that everything was downloaded by checking that each genome sequence is parseable and at least 1000 nts long**

problem_cases = []
for index in temp_df.index[:]:
    accession = temp_df.loc[index]['RefSeq accession number']
    if os.path.exists(fasta_dir + '{}.fasta'.format(accession)):
        record = SeqIO.read(fasta_dir + '{}.fasta'.format(accession), format='fasta')
        if len(record.seq) < 1000:
            problem_cases.append(accession)
print(len(problem_cases))

# Write six-frame translations for every fasta file in the dataset

Note that this code will fail on phages with non-standard genetic codes!

for fasta_file in glob.glob(fasta_dir+'*.fasta')[:]:
    print(fasta_file)
    bacphlip.six_frame_translate(fasta_file, fasta_file.replace('.fasta', '.fasta.6frame'),\
                                force_overwrite=True)

**fin.**
