==> Code/3-hmmer_time.ipynb <==
# Notes:

This entire file assumes that you have HMMER > 3.x installed on your computer. You may need to troubleshoot this file a bit if HMMER (specifically hmmbuild and hmmsearch) are not in the system path.

# Imports

import subprocess
import glob

# Run `hmmbuild` to build `hmmer` files from multiple sequence alignments

This takes a few seconds per domain but given the ~ small number of domains it should run fairly quickly (~5-10 minutes)

for infile in glob.glob('../Data/protein_domain_data/domain_alignments_and_hmms/*.afa'):
    print(infile)
    outfile = infile.replace('.afa', '.hmm')
    subprocess.run(args=['hmmbuild', outfile, infile], check=True)

**And concatenate all the .hmm files together into one large file**

all_files = glob.glob('../Data/protein_domain_data/domain_alignments_and_hmms/*.hmm')
with open('../Data/protein_domain_data/all_current_models.hmm', 'w') as outfile:
    for fname in all_files:
        with open(fname, 'r') as infile:
            outfile.write(infile.read())

# Next, run `hmmsearch` to search each genome file for each hmm model built above

This takes a few seconds for each of the >1,000 files so just note that this may take an hour or two to run depending on your computer

hmmfile = '../Data/protein_domain_data/all_current_models.hmm'

for faa_file in glob.glob('../Data/phage_data_nmicro2017/phage_fasta_files/*.6frame'):
    print(faa_file)
    outfile_path = faa_file.replace('/phage_fasta_files/', '/hmmsearch_out/').replace('.6frame', '.hmmsearch')
    with open(outfile_path, 'w') as outfile:
        subprocess.run(args=['hmmsearch', hmmfile, faa_file], check=True, stdout=outfile)
